refactor: 서비스로 Request 객체 넘기는 것은 안티패턴이므로 제거
