+++ src/main/java/umc/cockple/demo/domain/chat/service/ChatWebSocketService.java
@@ -28,33 +28,21 @@ public class ChatWebSocketService {
 
     private final ChatConverter chatConverter;
 
-    public WebSocketMessageDTO.Response sendMessage(WebSocketMessageDTO.Request request, Long senderId) {
-        log.info("메시지 전송 시작 - 채팅방: {}, 발신자: {}", request.chatRoomId(), senderId);
+    public WebSocketMessageDTO.Response sendMessage(Long chatRoomId, String content, Long senderId) {
+        log.info("메시지 전송 시작 - 채팅방: {}, 발신자: {}", chatRoomId, senderId);
 
-        validateRequest(request);
-
-        ChatRoom chatRoom = findChatRoomOrThrow(request.chatRoomId());
+        ChatRoom chatRoom = findChatRoomOrThrow(chatRoomId);
         Member sender = findMemberOrThrow(senderId);
 
         validateChatRoomMember(chatRoom, sender);
 
         // TODO: 다양한 타입의 텍스트 전송가능하도록 변경해야 함
-        ChatMessage chatMessage = ChatMessage.create(chatRoom, sender, request.content(), MessageType.TEXT);
+        ChatMessage chatMessage = ChatMessage.create(chatRoom, sender, content, MessageType.TEXT);
         ChatMessage savedMessage = chatMessageRepository.save(chatMessage);
 
         log.info("메시지 저장 완료 - 메시지 ID: {}", savedMessage.getId());
 
-        return chatConverter.toSendMessageResponse(request, savedMessage, sender);
-    }
-
-    private void validateRequest(WebSocketMessageDTO.Request request) {
-        if(request.chatRoomId() == null) {
-            throw new ChatException(ChatErrorCode.CHATROOM_ID_NECESSARY);
-        }
-
-        if(request.content() == null || request.content().trim().isEmpty()) {
-            throw new ChatException(ChatErrorCode.CONTENT_NECESSARY);
-        }
+        return chatConverter.toSendMessageResponse(chatRoomId, content, savedMessage, sender);
     }
 
     private void validateChatRoomMember(ChatRoom chatRoom, Member member) {
